% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_group_means.R
\name{compute_group_means}
\alias{compute_group_means}
\title{Compute replicate means for control and experiment groups}
\usage{
compute_group_means(df, control_name, experiment_name, metric = "IRratio")
}
\arguments{
\item{df}{Data frame produced by \code{read_irdir()} or user provied data frame(with replicate IRratio columns).}

\item{control_name}{Prefix for control sample (e.g., "HEK293T_scramble").}

\item{experiment_name}{Prefix for experiment sample (e.g., "HEK293T_snaR").}

\item{metric}{Which metric to average (default = "IRratio").}
}
\value{
A data frame with two new columns:
\itemize{
\item \verb{\{control_name\}_\{metric\}_avg}
\item \verb{\{experiment_name\}_\{metric\}_avg}
\item remove introns with IR ratio of minimum = 0 or maximum = 1  in either group
}
}
\description{
Given a merged IRFinder data frame or a user defined data frame, compute mean IRratio across replicates
for both control and experiment groups. The function looks for all replicate
columns matching \verb{\{sample_name\}_IRratio_repX}.
}
\examples{
\dontrun{
df <- compute_group_means(df, control_name = "HEK293T_scramble",
                               experiment_name = "HEK293T_snaR")
head(df)
}
}
