% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lfc_z.R
\name{compute_lfc_z}
\alias{compute_lfc_z}
\title{Compute log2 fold change (experiment over control) and transform into Z-score}
\usage{
compute_lfc_z(
  df,
  control_name,
  experiment_name,
  new_name,
  metric = "IRratio",
  pseudocount = 1e-06
)
}
\arguments{
\item{df}{Data frame (typically output of \code{compute_group_means()}).}

\item{control_name}{Character prefix of the control sample
(e.g., "HEK293T_scramble").}

\item{experiment_name}{Character prefix of the experiment sample
(e.g., "HEK293T_snaR").}

\item{new_name}{Character prefix of the new name
(e.g., "snaR").}

\item{metric}{Which averaged metric to use; default \code{"IRratio"}.
The function expects columns named \verb{\{sample_name\}_\{metric\}_avg}.}

\item{pseudocount}{Small numeric added to numerator and denominator to avoid
log of zero (default \code{1e-6}).}
}
\value{
The input data frame with two new columns:
\itemize{
\item \verb{\{new_name\}_\{metric\}_lfc}
\item \verb{\{new_name\}_\{metric\}_z}
}
}
\description{
Given a merged IRFinder table (with replicate means already computed),
calculate log2 fold change for \verb{\{experiment_name\}_\{metric\}_avg} over
\verb{\{control_name\}_\{metric\}_avg}, and add a standardized Z-score of that LFC.
}
\examples{
\dontrun{
df2 <- compute_lfc_z(
  df,
  experiment_name = "HEK293T_snaR",
  control_name    = "HEK293T_scramble",
  new_name = "snaR",
  metric = "IRratio"
)
}
}
